mBud/requirements.txt
apiai==1.2.3
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.1.2
dj-database-url==0.4.2
Django==1.11.7
django-tastypie==0.14.0
djangorestframework==3.7.3
enum34==1.1.6
FormEncode==1.3.1
google-api-python-client==1.6.4
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
IMDbPY==5.1.1
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.10
jupyter-client==5.1.0
jupyter-core==4.3.0
lxml==4.1.0
MarkupSafe==1.0
numpy==1.13.3
oauth2client==4.1.2
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
psycopg2==2.7.3.2
PyDispatcher==2.0.5
Pygments==2.2.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.3
pyzmq==16.0.2
requests==2.18.4
rsa==3.4.2
scandir==1.6
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.1.14
sqlalchemy-migrate==0.11.0
SQLObject==3.4.0
sqlparse==0.2.4
Tempita==0.5.2
tornado==4.5.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
whitenoise==3.3.1
